+++ consensus/hotstuff_leader.go
@@ -2,6 +2,7 @@ package consensus
 
 import (
 	"encoding/hex"
+	"github.com/pokt-network/pocket/shared/modules"
 	"unsafe"
 
 	consensusTelemetry "github.com/pokt-network/pocket/consensus/telemetry"
@@ -53,22 +54,27 @@ func (handler *HotstuffLeaderMessageHandler) HandleNewRoundMessage(m *ConsensusM
 	// TODO: Add more unit tests for these checks...
 	if m.shouldPrepareNewBlock(highPrepareQC) {
 		// Leader prepares a new block if `highPrepareQC` is not applicable
-		block, err := m.prepareAndApplyBlock()
+		block, txResults, err := m.prepareAndApplyBlock()
 		if err != nil {
 			m.nodeLogError(typesCons.ErrPrepareBlock.Error(), err)
 			m.paceMaker.InterruptRound()
 			return
 		}
 		m.Block = block
+		m.TxResults = txResults
 	} else {
 		// DISCUSS: Do we need to call `validateProposal` here?
 		// Leader acts like a replica if `highPrepareQC` is not `nil`
-		if err := m.applyBlock(highPrepareQC.Block); err != nil {
+		// DISCUSS: it seems like if the block was from the past round, it might already have
+		// been 'Applied'
+		txResults, err := m.applyBlock(highPrepareQC.Block)
+		if err != nil {
 			m.nodeLogError(typesCons.ErrApplyBlock.Error(), err)
 			m.paceMaker.InterruptRound()
 			return
 		}
 		m.Block = highPrepareQC.Block
+		m.TxResults = txResults
 	}
 
 	m.Step = Prepare
@@ -322,9 +328,9 @@ func (m *ConsensusModule) tempIndexHotstuffMessage(msg *typesCons.HotstuffMessag
 
 // This is a helper function intended to be called by a leader/validator during a view change
 // to prepare a new block that is applied to the new underlying context.
-func (m *ConsensusModule) prepareAndApplyBlock() (*typesCons.Block, error) {
+func (m *ConsensusModule) prepareAndApplyBlock() (*typesCons.Block, []modules.TxResult, error) {
 	if m.isReplica() {
-		return nil, typesCons.ErrReplicaPrepareBlock
+		return nil, nil, typesCons.ErrReplicaPrepareBlock
 	}
 
 	// TECHDEBT: Retrieve this from consensus consensus config
@@ -334,16 +340,16 @@ func (m *ConsensusModule) prepareAndApplyBlock() (*typesCons.Block, error) {
 	lastByzValidators := make([][]byte, 0)
 
 	// Reap the mempool for transactions to be applied in this block
-	txs, err := m.utilityContext.GetProposalTransactions(m.privateKey.Address(), maxTxBytes, lastByzValidators)
+	txs, _, err := m.utilityContext.GetProposalTransactions(m.privateKey.Address(), maxTxBytes, lastByzValidators)
 	if err != nil {
-		return nil, err
+		return nil, nil, err
 	}
 
 	// OPTIMIZE: Determine if we can avoid the `ApplyBlock` call here
 	// Apply all the transactions in the block
-	appHash, err := m.utilityContext.ApplyBlock(int64(m.Height), m.privateKey.Address(), txs, lastByzValidators)
+	appHash, txResults, err := m.utilityContext.ApplyBlock(int64(m.Height), m.privateKey.Address(), txs, lastByzValidators)
 	if err != nil {
-		return nil, err
+		return nil, nil, err
 	}
 
 	// Construct the block
@@ -360,7 +366,7 @@ func (m *ConsensusModule) prepareAndApplyBlock() (*typesCons.Block, error) {
 		Transactions: txs,
 	}
 
-	return block, nil
+	return block, txResults, nil
 }
 
 // Return true if this node, the leader, should prepare a new block