+++ consensus/module.go
@@ -44,10 +44,11 @@ type ConsensusModule struct {
 	m sync.RWMutex
 
 	// Hotstuff
-	Height uint64
-	Round  uint64
-	Step   typesCons.HotstuffStep
-	Block  *typesCons.Block // The current block being proposed / voted on; it has not been committed to finality
+	Height    uint64
+	Round     uint64
+	Step      typesCons.HotstuffStep
+	Block     *typesCons.Block   // The current block being proposed / voted on; it has not been committed to finality
+	TxResults []modules.TxResult // The current block applied transaction results / voted on; it has not been committed to finality
 
 	highPrepareQC *typesCons.QuorumCertificate // Highest QC for which replica voted PRECOMMIT
 	lockedQC      *typesCons.QuorumCertificate // Highest QC for which replica voted COMMIT