+++ utility/test/module_test.go
@@ -1,6 +1,7 @@
 package test
 
 import (
+	"encoding/hex"
 	"encoding/json"
 	"io/ioutil"
 	"log"
@@ -31,6 +32,7 @@ var (
 	defaultNonceString         = utilTypes.BigIntToString(test_artifacts.DefaultAccountAmount)
 	defaultSendAmountString    = utilTypes.BigIntToString(defaultSendAmount)
 	testSchema                 = "test_schema"
+	testMessageSendType        = "MessageSend"
 )
 
 var actorTypes = []utilTypes.ActorType{
@@ -128,5 +130,21 @@ func createTestingGenesisAndConfigFiles(cfg modules.Config, genesisState modules
 	}
 	if err := ioutil.WriteFile(testingConfigFilePath, configFileBz, 0777); err != nil {
 		log.Fatal(err)
+	}
+}
+
+func requireValidTestingTxResults(t *testing.T, tx *utilTypes.Transaction, txResults []modules.TxResult) {
+	for _, txResult := range txResults {
+		msg, err := tx.GetMessage()
+		sendMsg, ok := msg.(*utilTypes.MessageSend)
+		require.True(t, ok)
+		require.NoError(t, err)
+		require.Equal(t, int32(0), txResult.GetResultCode())
+		require.Equal(t, "", txResult.GetError())
+		require.Equal(t, testMessageSendType, txResult.GetMessageType())
+		require.Equal(t, int32(0), txResult.GetIndex())
+		require.Equal(t, int64(0), txResult.GetHeight())
+		require.Equal(t, hex.EncodeToString(sendMsg.ToAddress), txResult.GetRecipientAddr())
+		require.Equal(t, hex.EncodeToString(sendMsg.FromAddress), txResult.GetSignerAddr())
 	}
 }