+++ persistence/docs/CHANGELOG.md
@@ -6,6 +6,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.0.0.7] - 2022-10-12
+
+- Added `TxIndexer` sub-package (previously in Utility Module)
+- Added `TxIndexer` to both `PersistenceModule` and `PersistenceContext`
+- Implemented `TransactionExists` and `StoreTransaction`
+
 ## [0.0.0.6] - 2022-09-30
 
 - Removed no-op `DeleteActor` code